+++ InMemorySharding-node/src/main/java/com/soliatrdj9/imsd/service/serviceManager/service/ServiceManager.java
@@ -14,7 +14,7 @@ public interface ServiceManager {
 	
 	public void stopService();
 	
-	public void scaleOut(Float criteriaHashSeed) throws ExceptionStartNodeInShardingModeFailure, ExceptionStartNodeInShardingModeConflict;
+	public void scaleOut(Double criteriaHashSeed) throws ExceptionStartNodeInShardingModeFailure, ExceptionStartNodeInShardingModeConflict;
 	
 	public void scaleIn();
